minor cli reorg
